# Implemented proto2's unfortunate enum behavior