Proto2 puts enums into the unknown field set at parse time if the parsed value was not declared in the enum.

For singular fields, it is bad. For repeated fields, it is worse (the enum gets reordered, as illustrated in the test case).

This grows the code size and regresses performance a bit. (The performance regression would be worse, but I optimized it to use a bit test for the common case of enumerators between 0 and 63.)

It's unfortunate to have to spend code size and perf on this "feature", when history has shown it to be a mistake. It raises the question "can we just not do this?" At the moment I've implemented this because existing Python unit tests verify this behavior. But is there any chance we could break with the past here?

In particular, note that this does not really affect the wire. The data is preserved and reserialized either way (albeit in scrambled order for repeated fields), it's just a question of how it is presented through the message API to the application. So maybe there is hope here that a breaking change policy could make this behavior unnecessary.

```
name                                      old time/op  new time/op   delta
ArenaInitialBlockOneAlloc                 6.88ns ± 0%   6.37ns ± 1%   -7.34%  (p=0.000 n=12+12)
ArenaOneAlloc                             21.1ns ± 1%   20.8ns ± 1%   -1.34%  (p=0.000 n=12+11)
LoadAdsDescriptor_Proto2                  15.1ms ± 5%   15.1ms ± 3%     ~     (p=0.755 n=12+12)
LoadAdsDescriptor_Upb                     3.78ms ± 4%   3.78ms ± 3%     ~     (p=0.949 n=11+11)
LoadDescriptor_Proto2                      250µs ± 1%    245µs ± 3%   -1.63%  (p=0.000 n=12+12)
LoadDescriptor_Upb                        48.0µs ±17%   48.0µs ± 2%   -0.02%  (p=0.009 n=10+10)
Parse_Proto2<FileDesc,InitBlock,Copy>     17.0µs ± 1%   18.8µs ±13%  +10.96%  (p=0.000 n=10+12)
Parse_Proto2<FileDesc,NoArena,Copy>       31.7µs ±10%   29.7µs ± 9%   -6.48%  (p=0.003 n=12+10)
Parse_Proto2<FileDescSV,InitBlock,Alias>  17.1µs ±14%   17.3µs ± 3%   +1.18%  (p=0.002 n=10+10)
Parse_Proto2<FileDesc,UseArena,Copy>      20.2µs ± 3%   20.2µs ± 6%     ~     (p=0.766 n=9+11)
Parse_Upb_FileDesc<InitBlock,Alias>       8.54µs ± 0%   8.90µs ± 0%   +4.21%  (p=0.000 n=10+10)
Parse_Upb_FileDesc<InitBlock,Copy>        9.85µs ± 1%  10.68µs ± 0%   +8.43%  (p=0.000 n=11+11)
Parse_Upb_FileDesc<UseArena,Alias>        8.93µs ± 1%   9.36µs ± 1%   +4.73%  (p=0.000 n=12+11)
Parse_Upb_FileDesc<UseArena,Copy>         10.3µs ± 1%   11.1µs ± 1%   +7.86%  (p=0.000 n=10+12)
SerializeDescriptor_Proto2                6.43µs ±39%   5.56µs ± 6%     ~     (p=0.843 n=12+12)
SerializeDescriptor_Upb                   11.3µs ± 0%   11.6µs ± 0%   +2.92%  (p=0.000 n=11+11)


    FILE SIZE        VM SIZE    
 --------------  -------------- 
  +202% +1.71Ki  [ = ]       0    [Unmapped]
   +11% +1.35Ki   +11% +1.23Ki    upb/decode.c
    [NEW]    +531  [NEW]    +488    decode_enum_packed
    [NEW]    +326  [NEW]    +280    decode_checkenum_slow
    +3.2%    +320  +3.2%    +320    decode_msg
    [NEW]    +212  [NEW]    +168    decode_enum_toarray
    -2.1%      -5  [ = ]       0    decode_isdonefallback
  +2.6%    +676  +2.8%    +648    upb/def.c
    [NEW]    +610  [NEW]    +568    create_enumlayout
    +4.9%     +80  +5.0%     +80    create_enumdef
    +1.3%     +32  +1.3%     +32    resolve_msgdef
   -10.9%     -14  [ = ]       0    upb_oneofdef_field
    -0.3%     -16  -0.3%     -16    _upb_symtab_addfile
    -0.4%     -16  -0.4%     -16    create_msgdef
  +6.6%    +652  +3.8%    +128    bazel-out/k8-opt/bin/external/com_google_protobuf/src/google/protobuf/test_messages_proto2.upb.c
    [NEW]    +123  [NEW]     +16    protobuf_test_messages_proto2_TestAllTypesProto2_MapStringForeignEnumEntry_submsgs
    [NEW]    +117  [NEW]     +24    protobuf_test_messages_proto2_TestAllTypesProto2_NestedEnum_enuminit
    [NEW]    +114  [NEW]      +8    protobuf_test_messages_proto2_TestAllTypesProto2_MapStringNestedEnumEntry_submsgs
    [NEW]    +108  [NEW]      +8    protobuf_test_messages_proto2_TestAllTypesProto2_NestedEnum_enuminit_values
    [NEW]    +105  [NEW]     +24    protobuf_test_messages_proto2_ForeignEnumProto2_enuminit
    [NEW]     +53  [NEW]     +16    enums_layout
     +22%     +24  +100%     +24    src_google_protobuf_test_messages_proto2_proto_upb_file_layout
    +2.9%      +8  +4.2%      +8    protobuf_test_messages_proto2_TestAllTypesProto2_submsgs
  +9.1%    +647  +8.1%    +200    bazel-out/k8-opt/bin/external/com_google_protobuf/google/protobuf/descriptor.upb.c
    [NEW]    +118  [NEW]     +32    google_protobuf_MethodOptions_IdempotencyLevel_enuminit
    [NEW]    +100  [NEW]     +24    google_protobuf_FieldDescriptorProto_Label_enuminit
    [NEW]    +100  [NEW]     +24    google_protobuf_FieldOptions_JSType_enuminit
    [NEW]     +99  [NEW]     +24    google_protobuf_FieldDescriptorProto_Type_enuminit
    [NEW]     +98  [NEW]     +24    google_protobuf_FileOptions_OptimizeMode_enuminit
    [NEW]     +97  [NEW]     +24    google_protobuf_FieldOptions_CType_enuminit
     +35%     +24  +300%     +24    google_protobuf_FieldOptions_submsgs
     +19%     +16  +100%     +16    google_protobuf_FieldDescriptorProto_submsgs
    +9.2%      +8   +33%      +8    google_protobuf_MessageOptions_msginit
    +8.0%      +7  [ = ]       0    google_protobuf_DescriptorProto_msginit
    -5.9%      -6  [ = ]       0    google_protobuf_ServiceDescriptorProto_msginit
    -7.0%      -7  [ = ]       0    google_protobuf_FieldDescriptorProto_msginit
    -7.4%      -7  [ = ]       0    google_protobuf_SourceCodeInfo_msginit
  [ = ]       0   +43%     +24    [LOAD #3 [RW]]
  +1.5%     +16  +3.8%     +16    bazel-out/k8-opt/bin/external/com_google_protobuf/conformance/conformance.upb.c
     +16%     +16   +50%     +16    conformance_conformance_proto_upb_file_layout
  +5.7%     +16   +18%     +16    bazel-out/k8-opt/bin/external/com_google_protobuf/google/protobuf/any.upb.c
     +16%     +16   +50%     +16    google_protobuf_any_proto_upb_file_layout
  +5.5%     +16   +20%     +16    bazel-out/k8-opt/bin/external/com_google_protobuf/google/protobuf/duration.upb.c
     +17%     +16   +67%     +16    google_protobuf_duration_proto_upb_file_layout
  +5.5%     +16   +20%     +16    bazel-out/k8-opt/bin/external/com_google_protobuf/google/protobuf/field_mask.upb.c
     +16%     +16   +67%     +16    google_protobuf_field_mask_proto_upb_file_layout
  +1.4%     +16  +4.9%     +16    bazel-out/k8-opt/bin/external/com_google_protobuf/google/protobuf/struct.upb.c
     +16%     +16   +50%     +16    google_protobuf_struct_proto_upb_file_layout
  +5.3%     +16   +18%     +16    bazel-out/k8-opt/bin/external/com_google_protobuf/google/protobuf/timestamp.upb.c
     +17%     +16   +67%     +16    google_protobuf_timestamp_proto_upb_file_layout
  +1.0%     +16  +3.5%     +16    bazel-out/k8-opt/bin/external/com_google_protobuf/google/protobuf/wrappers.upb.c
     +17%     +16   +67%     +16    google_protobuf_wrappers_proto_upb_file_layout
  +0.2%     +16  +0.5%     +16    bazel-out/k8-opt/bin/external/com_google_protobuf/src/google/protobuf/test_messages_proto3.upb.c
     +14%     +16   +67%     +16    src_google_protobuf_test_messages_proto3_proto_upb_file_layout
  +0.4%     +13  [ = ]       0    tests/conformance_upb.c
    +6.3%     +13  [ = ]       0    main
  +0.1%     +10  [ = ]       0    upb/table.c
    +5.6%     +10  [ = ]       0    upb_inttable_done
  +0.1%      +8  +0.1%      +8    [section .rodata]
  [ = ]       0  [NEW]      +4    [section .bss]
    [ = ]       0  [NEW]      +4    completed.0
  -1.4%      -9  [ = ]       0    [section .strtab]
  [ = ]       0  -0.5%     -12    upb/upb.c
    [ = ]       0  [DEL]      -4    completed.0
    [ = ]       0 -50.0%      -8    upb_alloc_global
  +2.9% +5.13Ki  +1.7% +2.33Ki    TOTAL
```